# feat: allow converting otel trace ids to datadog values