--- a/lib/logger_json/formatters/datadog_logger.ex
+++ b/lib/logger_json/formatters/datadog_logger.ex
@@ -73,8 +73,8 @@ defmodule LoggerJSON.Formatters.DatadogLogger do
   # https://docs.datadoghq.com/tracing/faq/why-cant-i-see-my-correlated-logs-in-the-trace-id-panel/?tab=jsonlogs
   defp convert_tracing_keys(output, md) do
     fields = %{
-      span_id: ["dd.span_id", &(&1)],
-      trace_id: ["dd.trace_id", &(&1)],
+      span_id: ["dd.span_id", & &1],
+      trace_id: ["dd.trace_id", & &1],
       otel_span_id: ["dd.span_id", &convert_otel_field/1],
       otel_trace_id: ["dd.trace_id", &convert_otel_field/1]
     }