--- a/test/unit/logger_json_datadog_test.exs
+++ b/test/unit/logger_json_datadog_test.exs
@@ -226,7 +226,12 @@ defmodule LoggerJSONDatadogTest do
 
     test "convert otel_trace_id/otel_span_id to expected datadog keys" do
       Logger.configure_backend(LoggerJSON, metadata: :all)
-      Logger.metadata(otel_trace_id: <<98, 56, 49, 48, 100, 98, 97, 50, 57, 56, 48, 51, 101, 101, 54, 49, 101, 55, 99, 55, 49, 102, 102, 48, 99, 50, 99, 57, 53, 97, 57, 100>>)
+
+      Logger.metadata(
+        otel_trace_id:
+          <<98, 56, 49, 48, 100, 98, 97, 50, 57, 56, 48, 51, 101, 101, 54, 49, 101, 55, 99, 55, 49, 102, 102, 48, 99,
+            50, 99, 57, 53, 97, 57, 100>>
+      )
 
       log =
         fn -> Logger.debug("hello") end