--- a/test/unit/logger_json_datadog_test.exs
+++ b/test/unit/logger_json_datadog_test.exs
@@ -224,7 +224,7 @@ defmodule LoggerJSONDatadogTest do
       assert Map.has_key?(log, "span_id") == false
     end
 
-    test "convert otel_trace_id/otel_span_id to expected datadog keys" do
+    test "convert otel_trace_id/otel_span_id binary to expected datadog keys" do
       Logger.configure_backend(LoggerJSON, metadata: :all)
 
       Logger.metadata(
@@ -239,6 +239,48 @@ defmodule LoggerJSONDatadogTest do
         |> Jason.decode!()
 
       assert %{"dd.trace_id" => "16701352862047361693"} = log
+      assert Map.has_key?(log, "trace_id") == false
+      assert Map.has_key?(log, "span_id") == false
+    end
+
+    test "convert otel_trace_id/otel_span_id charlist to expected datadog keys" do
+      Logger.configure_backend(LoggerJSON, metadata: :all)
+      Logger.metadata(otel_trace_id: 'b810dba29803ee61e7c71ff0c2c95a9d')
+
+      log =
+        fn -> Logger.debug("hello") end
+        |> capture_log()
+        |> Jason.decode!()
+
+      assert %{"dd.trace_id" => "16701352862047361693"} = log
+      assert Map.has_key?(log, "trace_id") == false
+      assert Map.has_key?(log, "span_id") == false
+    end
+
+    test "convert otel_trace_id/otel_span_id string to expected datadog keys" do
+      Logger.configure_backend(LoggerJSON, metadata: :all)
+      Logger.metadata(otel_trace_id: "e7c71ff0c2c95a9d")
+
+      log =
+        fn -> Logger.debug("hello") end
+        |> capture_log()
+        |> Jason.decode!()
+
+      assert %{"dd.trace_id" => "16701352862047361693"} = log
+      assert Map.has_key?(log, "trace_id") == false
+      assert Map.has_key?(log, "span_id") == false
+    end
+
+    test "does not error on incorrect otel_trace_id/otel_span_id metadata" do
+      Logger.configure_backend(LoggerJSON, metadata: :all)
+      Logger.metadata(otel_trace_id: {:noop})
+
+      log =
+        fn -> Logger.debug("hello") end
+        |> capture_log()
+        |> Jason.decode!()
+
+      assert %{"dd.trace_id" => ""} = log
       assert Map.has_key?(log, "trace_id") == false
       assert Map.has_key?(log, "span_id") == false
     end