This is to allow Open Telemetry trace ids to be converted into Datadog format so logs and traces can be linked together. Datadog has a [page on how to do this](https://docs.datadoghq.com/tracing/other_telemetry/connect_logs_and_traces/opentelemetry/?tab=go). Most of the code was stollen from the golang example, and then the test values were stollen from known good versions based off that go code. You'll notice that the code is very defensive in case some bad value gets there and starts breaking things. ~I'll also mention that this is probably not the most optimized version. I'm not great with binary pattern matching so improvements welcome.~

Currently this is what would show up out of the box in Datadog with https://github.com/open-telemetry/opentelemetry-erlang setup.
![image](https://user-images.githubusercontent.com/3385679/192855067-edf40cb0-9e4b-47bf-b525-1d06b889e63a.png)

**Note** I'm going to be running this locally for a bit on a project in production to ensure nothing breaks and works as intended.