@btkostner 2022-09-28
I can confirm this is working in live production code now. I will continue to let it run in case something creeps up

![image](https://user-images.githubusercontent.com/3385679/192860787-11a10d49-7ee5-4372-b9ed-2ac57cd35dc6.png)