@btkostner 2022-09-28
Alright, after some function optimization one of my coworkers came up with a better solution. Benchmarks:

```
Name                      ips        average  deviation         median         99th %
new long string        1.62 M        0.62 μs  ±3721.34%        0.55 μs        1.07 μs
old long string        0.38 M        2.66 μs   ±602.14%        2.49 μs        3.62 μs

Name                       ips        average  deviation         median         99th %
new short string        1.79 M        0.56 μs  ±2125.00%        0.51 μs        0.90 μs
old short string        0.62 M        1.61 μs  ±1400.36%        1.49 μs        2.20 μs
```

```
Name                     Memory usage
new integer              0.0313 KB
old integer              3.70 KB - 118.50x memory usage +3.67 KB

new short string         0.26 KB
old short string         2.93 KB - 11.36x memory usage +2.67 KB

new long string          0.40 KB
old long string          6.02 KB - 15.12x memory usage +5.63 KB
```