@btkostner 2022-09-28
Some how missed charlist encoding not working. Should be fixed now. Working again on our side.