@AndrewDryga 2022-10-06
@btkostner thank you for another great PR ❤️. I'm moving between houses now and will be able to look more closely/merge it in a day or two, sorry for the delay.

Also, I got an idea I decided to discuss with you. I feel like change like this one can be generalized: with the new commit that added `formatter_state` we can actually make a configuration option (something like `metadata_mappers: [span_id: MyModule.otel_span_id/1]` but I bet there is a better name for it) and allow users to define custom functions that will process some of the keys. The function can receive a value and recturn a tuple `{new_key, new_value}`.

This will ensure that no matter what use case you have - you can always integrate OTEL or anything else that writes keys to the Logger metadata and you can control how it's encoded and under which key is stored. 

I'm thinking to implement such refactoring after we merge this PR. What do you think?