@btkostner 2022-10-06
No rush on the review. Take your time.

As for something like `metadata_mappers`, that sounds like a great idea. It should keep the logger pretty easy to extend for everyone's use case. If we find people are using a common function we could include it this library under a util module or something for ease of use. It could also cover the "filter params" use case.